custom_requirements.txt
aiohttp==3.7.4.post0
alembic==1.5.4
apache-airflow==2.0.1
apache-airflow-providers-ftp==1.0.1
apache-airflow-providers-http==1.1.0
apache-airflow-providers-imap==1.0.1
apache-airflow-providers-sqlite==1.0.1
apispec==3.3.2
argcomplete==1.12.2
asn1crypto==1.4.0
async-timeout==3.0.1
attrs==20.3.0
azure-common==1.1.27
azure-core==1.15.0
azure-storage-blob==12.8.1
Babel==2.9.0
base58==2.1.0
bitarray==1.2.2
boto3==1.17.92
botocore==1.20.92
cached-property==1.5.2
cachetools==4.2.2
cattrs==1.2.0
certifi==2020.12.5
cffi==1.14.5
chardet==3.0.4
click==7.1.2
clickclick==20.10.2
colorama==0.4.4
colorlog==4.7.2
commonmark==0.9.1
connexion==2.7.0
croniter==0.3.37
cryptography==3.4.5
cytoolz==0.11.0
defusedxml==0.6.0
dill==0.3.3
discord-notify==1.0.0
dnspython==1.16.0
docutils==0.16
email-validator==1.1.2
eth-abi==2.1.1
eth-account==0.5.4
eth-hash==0.3.1
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.2.1
eth-typing==2.2.2
eth-utils==1.10.0
Flask==1.1.2
Flask-AppBuilder==3.1.1
Flask-Babel==1.0.0
Flask-Caching==1.9.0
Flask-JWT-Extended==3.25.1
Flask-Login==0.4.1
Flask-OpenID==1.2.5
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
google-api-core==1.30.0
google-auth==1.31.0
google-cloud-bigquery==2.20.0
google-cloud-core==1.6.0
google-crc32c==1.1.2
google-resumable-media==1.3.0
googleapis-common-protos==1.53.0
graphviz==0.16
grpcio==1.38.0
gunicorn==19.10.0
hexbytes==0.2.1
idna==2.10
importlib-metadata==1.7.0
importlib-resources==1.5.0
inflection==0.5.1
ipfshttpclient==0.7.0
iso8601==0.1.14
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.3
jmespath==0.10.0
jsonschema==3.2.0
lazy-object-proxy==1.4.3
lockfile==0.12.2
lru-dict==1.1.7
Mako==1.1.4
Markdown==3.3.3
MarkupSafe==1.1.1
marshmallow==3.10.0
marshmallow-enum==1.5.1
marshmallow-oneofschema==2.1.0
marshmallow-sqlalchemy==0.23.1
msrest==0.6.21
multiaddr==0.0.9
multidict==5.1.0
natsort==7.1.1
netaddr==0.8.0
numpy==1.20.1
oauthlib==3.1.1
openapi-spec-validator==0.2.9
oscrypto==1.2.1
packaging==20.9
pandas==1.2.2
parsimonious==0.8.1
pendulum==2.1.2
prison==0.1.3
proto-plus==1.18.1
protobuf==3.17.3
psutil==5.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.10.1
pycryptodomex==3.10.1
Pygments==2.8.0
PyJWT==1.7.1
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-daemon==2.2.4
python-dateutil==2.8.1
python-dotenv==0.17.1
python-editor==1.0.4
python-nvd3==0.15.0
python-slugify==4.0.1
python3-openid==3.2.0
pytz==2020.5
pytzdata==2020.1
PyYAML==5.4.1
requests==2.23.0
requests-oauthlib==1.3.0
rich==9.2.0
rlp==2.0.1
rsa==4.7.2
s3transfer==0.4.2
setproctitle==1.2.2
six==1.15.0
snowflake-connector-python==2.4.4
SQLAlchemy==1.3.23
SQLAlchemy-JSONField==1.0.0
SQLAlchemy-Utils==0.36.8
swagger-ui-bundle==0.0.8
tabulate==0.8.7
tenacity==6.2.0
termcolor==1.1.0
text-unidecode==1.3
toolz==0.11.1
typing-extensions==3.7.4.3
unicodecsv==0.14.1
urllib3==1.25.11
varint==1.0.2
web3==5.20.0
websockets==8.1
Werkzeug==1.0.1
WTForms==2.3.3
yarl==1.6.3
zipp==3.4.0
selenium
tqdm
mailjet_rest
beautifulsoup4==4.10.0